Move checkPartitions() to DDLCommandTestUtils
